set up queryTraffic worker to be scheduled and will schedule firing of leave notification
