+++ server/utility/addTravel.js
@@ -1,14 +1,11 @@
 const TravelController = require('./../db/controllers/travelController');
 const requestPromise = require('request-promise');
-const queryTraffic = require('./../workers/queryTraffic');
-
 
 const origins = 'Hack Reactor, 944 Market St, San Francisco, CA 94102';
 const url = 'https://maps.googleapis.com/maps/api/distancematrix/json';
 
 
 const addTravel = function(event) {
-  console.log('inside addTravel', event)
   var options = {
     url,
     qs: {
@@ -40,8 +37,7 @@ const addTravel = function(event) {
   })
   .spread((travel, created) => {
     console.log('travel was created:', created);
-      // schedule job to query traffic, see workers/queryTraffic.js
-    queryTraffic(travel.dataValues.queryTime, travel.dataValues);
+    return travel;
   });
 };
